backend/requirements.txt
fastapi
uvicorn
httpx
playwright
cachetools
python-dotenv
pydantic
